+++ FrontEnd/src/App.tsx
@@ -14,7 +14,7 @@ import Rootpage from './pages/rootPage';
 import AdminInfomation from './pages/adminPage/info/adminInformation';
 import AdminRecord from './pages/adminPage/record/adminRecord';
 import AdminUser from './pages/adminPage/user/adminUser';
-import Community from './pages/communityPage/community';
+import Community from './pages/communityPage/Community';
 import Login from './pages/userPages/login';
 import Signup from './pages/userPages/signup';
 import Users from './pages/userPages/users';
